+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [2.5.0]
 
-- Added support for bulk operations: CreateMultipleRequest
+### Added
+
+- Added FileAttributeMetadata support to MetadataGenerator 
+- Added support for bulk operations: CreateMultipleRequest, UpdateMultipleRequest
 
 ## [2.4.2]